Bring back require $file-name;
Added an extra method to CompUnit::Repository for loading a module from a given
file name. The exact use cases for this are unclear, but it's speced and it
feels like an appropriate low level tool for extreme situations.

In any case it's quite separate from normal module loading, so it can be removed
easily in case we decide that it has to go.