CompUnit::Handle export-sub and export-package
We need to handle NQP and Perl 6 modules. So far we have standardized on he low
level way of NQP modules. The new architecture goes in the other direction and
tries to use high level objects as much as possible.

Loading the BOOTSTRAP is now done by separate code, since CompUnit::Handle is
not yet available there. But there we don't need most of the features do_import
provides, so code duplication can be kept quite minimal.